Refactoring stage II - move common functions from infinispan_controller to the separate packages Part of the #290 implementation
