+++ pkg/controller/infinispan/infinispan_controller.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"net/url"
 	"os"
 	"sort"
 	"strconv"
@@ -16,6 +15,7 @@ import (
 	consts "github.com/infinispan/infinispan-operator/pkg/controller/constants"
 	ispnutil "github.com/infinispan/infinispan-operator/pkg/controller/infinispan/util"
 	ispncom "github.com/infinispan/infinispan-operator/pkg/controller/utils/common"
+	"github.com/infinispan/infinispan-operator/pkg/controller/utils/infinispan/configuration"
 	"github.com/infinispan/infinispan-operator/version"
 	"gopkg.in/yaml.v2"
 	appsv1 "k8s.io/api/apps/v1"
@@ -27,8 +27,6 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/intstr"
-	restclient "k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
@@ -145,7 +143,8 @@ func (r *ReconcileInfinispan) Reconcile(request reconcile.Request) (reconcile.Re
 	// Check x-site configuration first.
 	// Must be done before creating any Infinispan resources,
 	// because remote site host:port combinations need to be injected into Infinispan.
-	xsite, err := r.computeXSite(infinispan, reqLogger)
+	xsite := &configuration.XSite{}
+	err = xsite.ComputeXSite(infinispan, kubernetes, r.scheme, reqLogger)
 	if err != nil {
 		reqLogger.Error(err, "Error in computeXSite", "Infinispan.Namespace")
 		return reconcile.Result{}, err
@@ -157,7 +156,7 @@ func (r *ReconcileInfinispan) Reconcile(request reconcile.Request) (reconcile.Re
 	if err != nil && errors.IsNotFound(err) {
 		reqLogger.Info("Configuring the StatefulSet")
 		// Populate EndpointEncryption if serving cert service is available
-		if getServingCertsMode(kubernetes) == "openshift.io" {
+		if kubernetes.GetServingCertsMode() == "openshift.io" {
 			reqLogger.Info("Serving certificate service present. Configuring into CRD")
 			requeue := false
 			ee := &infinispan.Spec.Security.EndpointEncryption
@@ -293,7 +292,7 @@ func (r *ReconcileInfinispan) Reconcile(request reconcile.Request) (reconcile.Re
 	} else if err != nil {
 		reqLogger.Error(err, "failed to get StatefulSet")
 		return reconcile.Result{}, err
-	} else if upgradeNeeded(infinispan, reqLogger) {
+	} else if infinispan.IsUpgradeNeeded(reqLogger) {
 		reqLogger.Info("Upgrade needed")
 		err = r.destroyResources(infinispan)
 		if err != nil {
@@ -302,7 +301,7 @@ func (r *ReconcileInfinispan) Reconcile(request reconcile.Request) (reconcile.Re
 		}
 
 		infinispan.SetCondition("upgrade", "False", "")
-		r.client.Status().Update(context.TODO(), infinispan)
+		err = r.client.Status().Update(context.TODO(), infinispan)
 		if err != nil {
 			reqLogger.Error(err, "failed to update Infinispan status")
 			return reconcile.Result{}, err
@@ -311,7 +310,7 @@ func (r *ReconcileInfinispan) Reconcile(request reconcile.Request) (reconcile.Re
 		reqLogger.Info("removed Infinispan resources, force an upgrade now", "replicasWantedAtRestart", infinispan.Status.ReplicasWantedAtRestart)
 
 		infinispan.Spec.Replicas = infinispan.Status.ReplicasWantedAtRestart
-		r.client.Update(context.TODO(), infinispan)
+		err = r.client.Update(context.TODO(), infinispan)
 		if err != nil {
 			reqLogger.Error(err, "failed to update Infinispan Spec")
 			return reconcile.Result{}, err
@@ -446,8 +445,8 @@ func (r *ReconcileInfinispan) Reconcile(request reconcile.Request) (reconcile.Re
 	}
 
 	// If upgrade required, do not process any further and handle upgrades
-	if isUpgradeCondition(infinispan) {
-		reqLogger.Info("isUpgradeCondition")
+	if infinispan.IsUpgradeCondition() {
+		reqLogger.Info("IsUpgradeCondition")
 		return reconcile.Result{Requeue: true}, nil
 	}
 
@@ -725,41 +724,11 @@ func (r *ReconcileInfinispan) destroyResources(infinispan *infinispanv1.Infinisp
 	return nil
 }
 
-func upgradeNeeded(infinispan *infinispanv1.Infinispan, logger logr.Logger) bool {
-	if isUpgradeCondition(infinispan) {
-		stoppingCondition := infinispan.GetCondition("stopping")
-		if stoppingCondition != nil {
-			stoppingCompleted := *stoppingCondition == "False"
-			if stoppingCompleted {
-				if infinispan.Status.ReplicasWantedAtRestart > 0 {
-					logger.Info("graceful shutdown after upgrade completed, continue upgrade process")
-					return true
-				}
-				logger.Info("replicas to restart with not yet set, wait for graceful shutdown to complete")
-				return false
-			}
-			logger.Info("wait for graceful shutdown before update to complete")
-			return false
-		}
-	}
-
-	return false
-}
-
-func isUpgradeCondition(infinispan *infinispanv1.Infinispan) bool {
-	return isConditionTrue("upgrade", infinispan)
-}
-
-func isConditionTrue(name string, infinispan *infinispanv1.Infinispan) bool {
-	upgradeCondition := infinispan.GetCondition(name)
-	return upgradeCondition != nil && *upgradeCondition == "True"
-}
-
 func (r *ReconcileInfinispan) scheduleUpgradeIfNeeded(infinispan *infinispanv1.Infinispan, podList *corev1.PodList, logger logr.Logger) {
 	if len(podList.Items) == 0 {
 		return
 	}
-	if isUpgradeCondition(infinispan) {
+	if infinispan.IsUpgradeCondition() {
 		return
 	}
 
@@ -874,122 +843,12 @@ func createCacheServiceDefaultCache(podName string, infinispan *infinispanv1.Inf
 	return cluster.CreateCacheWithTemplate(consts.DefaultOperatorUser, pass, consts.DefaultCacheName, defaultCacheXML, podName, infinispan.Namespace, string(protocol))
 }
 
-func (r *ReconcileInfinispan) computeXSite(infinispan *infinispanv1.Infinispan, logger logr.Logger) (*ispnutil.XSite, error) {
-	if infinispan.HasSites() {
-		siteServiceName := infinispan.GetSiteServiceName()
-		siteService, err := r.getOrCreateSiteService(siteServiceName, infinispan, logger)
-		if err != nil {
-			logger.Error(err, "could not get or create site service")
-			return nil, err
-		}
-
-		localSiteHost, localSitePort, err := getLocalSiteServiceHostPort(siteService, infinispan, logger)
-		if err != nil {
-			logger.Error(err, "error retrieving local x-site service information")
-			return nil, err
-		}
-
-		if localSiteHost == "" {
-			msg := "local x-site service host not yet available"
-			logger.Info(msg)
-			return nil, fmt.Errorf(msg)
-		}
-
-		logger.Info("local site service",
-			"service name", siteServiceName,
-			"host", localSiteHost,
-			"port", localSitePort,
-		)
-
-		xsite := &ispnutil.XSite{
-			Address: localSiteHost,
-			Name:    infinispan.Spec.Service.Sites.Local.Name,
-			Port:    localSitePort,
-		}
-
-		remoteLocations := findRemoteLocations(xsite.Name, infinispan)
-		for _, remoteLocation := range remoteLocations {
-			err := appendRemoteLocation(xsite, infinispan, &remoteLocation, logger)
-			if err != nil {
-				return nil, err
-			}
-		}
-
-		logger.Info("x-site configured", "configuration", *xsite)
-		return xsite, nil
-	}
-
-	return nil, nil
-}
-
-func findRemoteLocations(localSiteName string, infinispan *infinispanv1.Infinispan) (remoteLocations []infinispanv1.InfinispanSiteLocationSpec) {
-	locations := infinispan.Spec.Service.Sites.Locations
-	for _, location := range locations {
-		if localSiteName != location.Name {
-			remoteLocations = append(remoteLocations, location)
-		}
-	}
-
-	return
-}
-
-func getLocalSiteServiceHostPort(service *corev1.Service, infinispan *infinispanv1.Infinispan, logger logr.Logger) (string, int32, error) {
-	switch serviceType := service.Spec.Type; serviceType {
-	case corev1.ServiceTypeNodePort:
-		externalIP, err := infinispan.FindNodePortExternalIP()
-		if err != nil {
-			return "", 0, err
-		}
-
-		// If configuring NodePort, expect external IPs to be configured
-		return externalIP, kubernetes.GetNodePort(service), nil
-	case corev1.ServiceTypeLoadBalancer:
-		return getLoadBalancerServiceHostPort(service, logger)
-	default:
-		return "", 0, fmt.Errorf("unsupported service type '%v'", serviceType)
-	}
-}
-
-func getLoadBalancerServiceHostPort(service *corev1.Service, logger logr.Logger) (string, int32, error) {
-	port := service.Spec.Ports[0].Port
-
-	// If configuring load balancer, look for external ingress
-	if len(service.Status.LoadBalancer.Ingress) > 0 {
-		ingress := service.Status.LoadBalancer.Ingress[0]
-		if ingress.IP != "" {
-			return ingress.IP, port, nil
-		}
-		if ingress.Hostname != "" {
-			// Resolve load balancer host
-			ip, err := ispncom.LookupHost(ingress.Hostname, logger)
-
-			// Load balancer gets created asynchronously,
-			// so it might take time for the status to be updated.
-			return ip, port, err
-		}
-	}
-
-	return "", port, nil
-}
-
 // deploymentForInfinispan returns an infinispan Deployment object
 func (r *ReconcileInfinispan) deploymentForInfinispan(m *infinispanv1.Infinispan, secret *corev1.Secret, configMap *corev1.ConfigMap) (*appsv1.StatefulSet, error) {
 	reqLogger := log.WithValues("Request.Namespace", m.Namespace, "Request.Name", m.Name)
 	lsPod := ispncom.LabelsResource(m.ObjectMeta.Name, "infinispan-pod")
-	// This field specifies the flavor of the
-	// Infinispan cluster. "" is plain community edition (vanilla)
-	var imageName string
-	if m.Spec.Image != "" {
-		imageName = m.Spec.Image
-	} else {
-		imageName = consts.DefaultImageName
-	}
 
-	memory := consts.DefaultMemorySize
-	if m.Spec.Container.Memory != "" {
-		memory = resource.MustParse(m.Spec.Container.Memory)
-	}
-
+	memory := resource.MustParse(m.Spec.Container.Memory)
 	cpuRequests, cpuLimits := m.GetCpuResources()
 
 	javaOptions, err := m.GetJavaOptions()
@@ -1020,10 +879,7 @@ func (r *ReconcileInfinispan) deploymentForInfinispan(m *infinispanv1.Infinispan
 		}
 	}
 	replicas := m.Spec.Replicas
-	protocolScheme := corev1.URISchemeHTTP
-	if m.GetEndpointScheme() != corev1.URISchemeHTTP {
-		protocolScheme = corev1.URISchemeHTTPS
-	}
+	protocolScheme := m.GetEndpointScheme()
 	dep := &appsv1.StatefulSet{
 
 		TypeMeta: metav1.TypeMeta{
@@ -1049,7 +905,7 @@ func (r *ReconcileInfinispan) deploymentForInfinispan(m *infinispanv1.Infinispan
 				},
 				Spec: corev1.PodSpec{
 					Containers: []corev1.Container{{
-						Image: imageName,
+						Image: m.Spec.Image,
 						Name:  "infinispan",
 						Env:   envVars,
 						LivenessProbe: &corev1.Probe{
@@ -1196,7 +1052,7 @@ func setupVolumesForEncryption(m *infinispanv1.Infinispan, dep *appsv1.StatefulS
 	}
 }
 
-func setupConfigForEncryption(m *infinispanv1.Infinispan, c *ispnutil.InfinispanConfiguration, client client.Client) error {
+func setupConfigForEncryption(m *infinispanv1.Infinispan, c *configuration.InfinispanConfiguration, client client.Client) error {
 	ee := m.Spec.Security.EndpointEncryption
 	if ee.Type == "service" {
 		if strings.Contains(ee.CertServiceName, "openshift.io") {
@@ -1237,12 +1093,12 @@ func setupConfigForEncryption(m *infinispanv1.Infinispan, c *ispnutil.Infinispan
 	return nil
 }
 
-func (r *ReconcileInfinispan) configMapForInfinispan(xsite *ispnutil.XSite, m *infinispanv1.Infinispan) (*corev1.ConfigMap, error) {
+func (r *ReconcileInfinispan) configMapForInfinispan(xsite *configuration.XSite, m *infinispanv1.Infinispan) (*corev1.ConfigMap, error) {
 	name := m.ObjectMeta.Name
 	namespace := m.ObjectMeta.Namespace
 
 	loggingCategories := m.CopyLoggingCategories()
-	config := ispnutil.CreateInfinispanConfiguration(name, xsite, loggingCategories, namespace)
+	config := xsite.CreateInfinispanConfiguration(name, loggingCategories, namespace)
 	err := setupConfigForEncryption(m, &config, r.client)
 	if err != nil {
 		return nil, err
@@ -1461,224 +1317,6 @@ func (r *ReconcileInfinispan) serviceExternal(m *infinispanv1.Infinispan) *corev
 	// Set Infinispan instance as the owner and controller
 	controllerutil.SetControllerReference(m, externalService, r.scheme)
 	return externalService
-}
-
-func (r *ReconcileInfinispan) getOrCreateSiteService(siteServiceName string, infinispan *infinispanv1.Infinispan, logger logr.Logger) (*corev1.Service, error) {
-	siteService := &corev1.Service{}
-	ns := types.NamespacedName{
-		Name:      siteServiceName,
-		Namespace: infinispan.ObjectMeta.Namespace,
-	}
-	err := r.client.Get(context.TODO(), ns, siteService)
-
-	if errors.IsNotFound(err) {
-		return r.createSiteService(siteServiceName, infinispan, logger)
-	}
-
-	if err != nil {
-		return nil, err
-	}
-
-	return siteService, nil
-}
-
-func (r *ReconcileInfinispan) createSiteService(siteServiceName string, infinispan *infinispanv1.Infinispan, logger logr.Logger) (*corev1.Service, error) {
-	lsPodSelector := ispncom.LabelsResource(infinispan.ObjectMeta.Name, "infinispan-pod")
-	lsService := ispncom.LabelsResource(infinispan.ObjectMeta.Name, "infinispan-service-xsite")
-	exposeSpec := infinispan.Spec.Service.Sites.Local.Expose
-	exposeSpec.Selector = lsPodSelector
-
-	switch exposeSpec.Type {
-	case corev1.ServiceTypeNodePort:
-		exposeSpec.Ports = []corev1.ServicePort{
-			{
-				Port:     7900,
-				NodePort: 32660,
-			},
-		}
-	case corev1.ServiceTypeLoadBalancer:
-		exposeSpec.Ports = []corev1.ServicePort{
-			{
-				Port: 7900,
-			},
-		}
-	}
-
-	logger.Info("create exposed site service", "configuration", exposeSpec)
-
-	siteService := corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      siteServiceName,
-			Namespace: infinispan.ObjectMeta.Namespace,
-			Annotations: map[string]string{
-				"service.beta.kubernetes.io/aws-load-balancer-backend-protocol": "tcp",
-			},
-			Labels: lsService,
-		},
-		Spec: exposeSpec,
-	}
-
-	// Set Infinispan instance as the owner and controller
-	controllerutil.SetControllerReference(infinispan, &siteService, r.scheme)
-
-	err := r.client.Create(context.TODO(), &siteService)
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return nil, err
-	}
-
-	return &siteService, nil
-}
-
-func appendRemoteLocation(xsite *ispnutil.XSite, infinispan *infinispanv1.Infinispan, remoteLocation *infinispanv1.InfinispanSiteLocationSpec, logger logr.Logger) error {
-	restConfig, err := GetRemoteSiteRESTConfig(infinispan, remoteLocation, logger)
-	if err != nil {
-		return err
-	}
-
-	remoteKubernetes, err := ispnutil.NewKubernetesFromConfig(restConfig)
-	if err != nil {
-		logger.Error(err, "could not connect to remote location URL", "URL", remoteLocation.URL)
-		return err
-	}
-
-	err = appendKubernetesRemoteLocation(xsite, infinispan, remoteLocation, remoteKubernetes, logger)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func GetRemoteSiteRESTConfig(infinispan *infinispanv1.Infinispan, location *infinispanv1.InfinispanSiteLocationSpec, logger logr.Logger) (*restclient.Config, error) {
-	backupSiteURL, err := url.Parse(location.URL)
-	if err != nil {
-		return nil, err
-	}
-
-	// Copy URL to make modify it for backup access
-	copyURL, err := url.Parse(backupSiteURL.String())
-	if err != nil {
-		return nil, err
-	}
-
-	// All remote sites locations are accessed via encrypted http
-	copyURL.Scheme = "https"
-
-	switch scheme := backupSiteURL.Scheme; scheme {
-	case "minikube":
-		return getMinikubeRESTConfig(copyURL.String(), location.SecretName, infinispan, logger)
-	case "openshift":
-		return getOpenShiftRESTConfig(copyURL.String(), location.SecretName, infinispan, logger)
-	default:
-		return nil, fmt.Errorf("backup site URL scheme '%s' not supported", scheme)
-	}
-}
-
-func getMinikubeRESTConfig(masterURL string, secretName string, infinispan *infinispanv1.Infinispan, logger logr.Logger) (*restclient.Config, error) {
-	logger.Info("connect to backup minikube cluster", "url", masterURL)
-
-	config, err := clientcmd.BuildConfigFromFlags(masterURL, "")
-	if err != nil {
-		logger.Error(err, "unable to create REST configuration", "master URL", masterURL)
-		return nil, err
-	}
-
-	secret, err := kubernetes.GetSecret(secretName, infinispan.ObjectMeta.Namespace)
-	if err != nil {
-		logger.Error(err, "unable to find Secret", "secret name", secretName)
-		return nil, err
-	}
-
-	config.CAData = secret.Data["certificate-authority"]
-	config.CertData = secret.Data["client-certificate"]
-	config.KeyData = secret.Data["client-key"]
-
-	return config, nil
-}
-
-func getOpenShiftRESTConfig(masterURL string, secretName string, infinispan *infinispanv1.Infinispan, logger logr.Logger) (*restclient.Config, error) {
-	config, err := clientcmd.BuildConfigFromFlags(masterURL, "")
-	if err != nil {
-		logger.Error(err, "unable to create REST configuration", "master URL", masterURL)
-		return nil, err
-	}
-
-	// Skip-tls for accessing other OpenShift clusters
-	config.Insecure = true
-
-	secret, err := kubernetes.GetSecret(secretName, infinispan.ObjectMeta.Namespace)
-	if err != nil {
-		logger.Error(err, "unable to find Secret", "secret name", secretName)
-		return nil, err
-	}
-
-	if token, ok := secret.Data["token"]; ok {
-		config.BearerToken = string(token)
-		return config, nil
-	}
-
-	return nil, fmt.Errorf("token required connect to OpenShift cluster")
-}
-
-func appendKubernetesRemoteLocation(xsite *ispnutil.XSite, infinispan *infinispanv1.Infinispan, remoteLocation *infinispanv1.InfinispanSiteLocationSpec, remoteKubernetes *ispnutil.Kubernetes, logger logr.Logger) error {
-	siteServiceName := infinispan.GetSiteServiceName()
-	namespacedName := types.NamespacedName{Name: siteServiceName, Namespace: infinispan.Namespace}
-	siteService := &corev1.Service{}
-	err := remoteKubernetes.Client.Get(context.TODO(), namespacedName, siteService)
-	if err != nil {
-		logger.Error(err, "could not find x-site service in remote cluster", "site service name", siteServiceName)
-		return err
-	}
-
-	host, port, err := getRemoteSiteServiceHostPort(siteService, remoteKubernetes, logger)
-	if err != nil {
-		logger.Error(err, "error retrieving remote x-site service information")
-		return err
-	}
-
-	if host == "" {
-		msg := "remote x-site service host not yet available"
-		logger.Info(msg)
-		return fmt.Errorf(msg)
-	}
-
-	logger.Info("remote site service",
-		"service name", siteServiceName,
-		"host", host,
-		"port", port,
-	)
-
-	backupSite := ispnutil.BackupSite{
-		Address: host,
-		Name:    remoteLocation.Name,
-		Port:    port,
-	}
-
-	xsite.Backups = append(xsite.Backups, backupSite)
-	return nil
-}
-
-func getRemoteSiteServiceHostPort(service *corev1.Service, remoteKubernetes *ispnutil.Kubernetes, logger logr.Logger) (string, int32, error) {
-	switch serviceType := service.Spec.Type; serviceType {
-	case corev1.ServiceTypeNodePort:
-		// If configuring NodePort, expect external IPs to be configured
-		return remoteKubernetes.PublicIP(), remoteKubernetes.GetNodePort(service), nil
-	case corev1.ServiceTypeLoadBalancer:
-		return getLoadBalancerServiceHostPort(service, logger)
-	default:
-		return "", 0, fmt.Errorf("unsupported service type '%v'", serviceType)
-	}
-}
-
-// getServingCertsMode returns a label that identify the kind of serving
-// certs service is available. Returns 'openshift.io' for service-ca on openshift
-func getServingCertsMode(remoteKubernetes *ispnutil.Kubernetes) string {
-	if remoteKubernetes.HasServiceCAsCRDResource() {
-		return "openshift.io"
-
-		// Code to check if other modes of serving TLS cert service is available
-		// can be added here
-	}
-	return ""
 }
 
 func (r *ReconcileInfinispan) finalizeInfinispan(reqLogger logr.Logger, ispn *infinispanv1.Infinispan) error {
