+++ test/e2e/utils/common.go
@@ -1,0 +1,29 @@
+package utils
+
+import (
+	"bufio"
+	"os"
+	"path/filepath"
+	"strings"
+
+	"k8s.io/apimachinery/pkg/util/yaml"
+)
+
+func GetYamlReaderFromFile(filename string) (*yaml.YAMLReader, error) {
+	absFileName := getAbsolutePath(filename)
+	f, err := os.Open(absFileName)
+	if err != nil {
+		return nil, err
+	}
+	return yaml.NewYAMLReader(bufio.NewReader(f)), nil
+}
+
+// Obtain the file absolute path given a relative path
+func getAbsolutePath(relativeFilePath string) string {
+	if !strings.HasPrefix(relativeFilePath, ".") {
+		return relativeFilePath
+	}
+	dir, _ := os.Getwd()
+	absPath, _ := filepath.Abs(dir + "/" + relativeFilePath)
+	return absPath
+}
